feat(refactor-mcts): Separate InvalidChoice class from ActionChoices
- Add comments for doxygen
