+++ Includes/Rosetta/Actions/Choices/InvalidChoice.hpp
@@ -1,0 +1,53 @@
+// Copyright (c) 2019 Chris Ohk, Youngjoong Kim, SeungHyun Jeon
+
+// We are making my contributions/submissions to this project solely in our
+// personal capacity and are not conveying any rights to any intellectual
+// property of any third parties.
+
+// It is based on peter1591's hearthstone-ai repository.
+// References: https://github.com/peter1591/hearthstone-ai
+
+#ifndef ROSETTASTONE_INVALID_ACTION_HPP
+#define ROSETTASTONE_INVALID_ACTION_HPP
+
+namespace RosettaStone
+{
+//!
+//! \brief ActionChoices::InvalidChoice class.
+//!
+//! This class indicates the choice is invalid.
+//! Note that all methods will throw exception.
+//!
+class InvalidChoice
+{
+ public:
+    //! Returns the action choice at \p idx.
+    //! \param idx The index of action choices.
+    //! \return The action choice at \p idx.
+    static size_t Get([[maybe_unused]] size_t idx);
+
+    //! Returns the flag indicates that action choices are empty.
+    //! \return The flag indicates that action choices are empty.
+    static bool IsEmpty();
+
+    //! Returns the number of action choices.
+    //! \return The number of action choices.
+    static size_t Size();
+
+    //! Initializes iterator variable.
+    static void Begin();
+
+    //! Returns the element that iterator variable points to.
+    //! \return The element that iterator variable points to.
+    static size_t Get();
+
+    //! Processes iterator to point to the next element.
+    static void StepNext();
+
+    //! Returns the flag indicates that iterator is in the end position.
+    //! \return the flag indicates that iterator is in the end position.
+    static bool IsEnd();
+};
+}  // namespace RosettaStone
+
+#endif  // ROSETTASTONE_INVALID_ACTION_HPP
